refactor(search): extract SCD + dimensional-role tools to _ScdAndRoleMixin
Moves the two heaviest live-DB classification tools out of ToolBox:

- _tool_detect_scd_pattern (~338 LOC) — slowly-changing dimension
  pattern detection over date/version columns.
- _tool_detect_dimensional_role (~152 LOC) — single-table /
  multi-table classification (fact / dimension / measure / staging).

Plus their three private helpers (_name_role_signal,
_count_column_shape, _classify_table_role) and 11 naming-pattern
constants (_SCD_*, _DIM_ROLE_NAMING, _MEASURE/_ID/_DESCRIPTIVE_NAME_PATTERNS).

The mixin reads self._live_db on the host ToolBox for the SQL samples
each tool needs. agent_tools.py: 4103 -> 3177 LOC (-926, -23%).
